smart-web-summarizer/requirements.txt

requests
beautifulsoup4
python-dotenv
openai
